Move `DisplayFolder` to `:legacy:ui:folder` module
