+++ feature/navigation/drawer/src/main/kotlin/app/k9mail/feature/navigation/drawer/legacy/FolderList.kt
@@ -1,6 +1,6 @@
 package app.k9mail.feature.navigation.drawer.legacy
 
-import app.k9mail.legacy.folder.DisplayFolder
+import app.k9mail.legacy.ui.folder.DisplayFolder
 
 data class FolderList(
     val unifiedInbox: DisplayUnifiedInbox?,
